put msgpool behind an interface
