+++ rpc/transport.go
@@ -51,13 +51,13 @@ type StreamTransport struct {
 	okay int32 // 1 = okay, 0 = failed, -1 = failing
 	err  error
 
-	msgPool *MsgPool
+	msgPool MessagePool
 }
 
 // Create a stream transport from a connected stream socket.  This is
 // the only valid way to initialize a StreamTransport.  You can
 // manually adjust MaxMsgSize after calling this function.
-func NewStreamTransport(c net.Conn, mp *MsgPool) *StreamTransport {
+func NewStreamTransport(c net.Conn, mp MessagePool) *StreamTransport {
 	return &StreamTransport{
 		MaxMsgSize: 0x100000,
 		Conn:       c,
